case.setup: Add --keep option
A convenient way to keep local changes to key generated files when
cleaning or resetting a case.

+++ scripts/lib/CIME/case/case_setup.py
@@ -67,7 +67,7 @@ def _build_usernl_files(case, model, comp):
                     safe_copy(model_nl, nlfile)
 
 ###############################################################################
-def _case_setup_impl(case, caseroot, clean=False, test_mode=False, reset=False):
+def _case_setup_impl(case, caseroot, clean=False, test_mode=False, reset=False, keep=None):
 ###############################################################################
     os.chdir(caseroot)
 
@@ -86,7 +86,7 @@ def _case_setup_impl(case, caseroot, clean=False, test_mode=False, reset=False):
         batch_script = get_batch_script_for_job(case.get_primary_job())
         files_to_clean = [batch_script, "env_mach_specific.xml", "Macros.make", "Macros.cmake"]
         for file_to_clean in files_to_clean:
-            if os.path.exists(file_to_clean):
+            if os.path.exists(file_to_clean) and not (keep and file_to_clean in keep):
                 os.remove(file_to_clean)
                 logger.info("Successfully cleaned {}".format(file_to_clean))
 
@@ -229,11 +229,11 @@ def _case_setup_impl(case, caseroot, clean=False, test_mode=False, reset=False):
         logger.info("You can now run './preview_run' to get more info on how your case will be run")
 
 ###############################################################################
-def case_setup(self, clean=False, test_mode=False, reset=False):
+def case_setup(self, clean=False, test_mode=False, reset=False, keep=None):
 ###############################################################################
     caseroot, casebaseid = self.get_value("CASEROOT"), self.get_value("CASEBASEID")
     phase = "setup.clean" if clean else "case.setup"
-    functor = lambda: _case_setup_impl(self, caseroot, clean, test_mode, reset)
+    functor = lambda: _case_setup_impl(self, caseroot, clean=clean, test_mode=test_mode, reset=reset, keep=keep)
 
     if self.get_value("TEST") and not test_mode:
         test_name = casebaseid if casebaseid is not None else self.get_value("CASE")
